バリデーションとエラーメッセージを分離
